Split agent tests into agent/tests/{support,document_loading,builder,supervision}.rs
