+++ crates/defra-agent/src/agent/tests/mod.rs
@@ -1,0 +1,4 @@
+mod support;
+mod document_loading;
+mod builder;
+mod supervision;
